requirements.txt
torch==1.13.0
torchvision==0.14.0
numpy==1.23.5
SimpleITK==2.2.0
nibabel==4.0.2
segmentation_models_pytorch==0.3.1
wandb==0.13.6
git+https://github.com/wolny/pytorch-3dunet.git@1.3.4
h5py~=3.7.0
pytorch-lightning==1.8.3.post1
matplotlib==3.6.2
volumentations-3D==1.0.4
opencv-python==4.7.0.68

